refactor(sdk): move RPC types to a separate dir
